Refactor ModelMetadata
Use getFieldHash() for getting the prop types as well, passing in convertToPropType.
